chore: restructure test files
