+++ tests/unit/askBirthday.test.ts
@@ -1,4 +1,4 @@
-import "@/testUtils/mocks/mockSlackApp";
+import "@/testUtils/unit/mockSlackApp";
 
 import type { Mock } from "vitest";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
@@ -8,7 +8,7 @@ import { handler } from "@/functions/events/askBirthday";
 import { constructAskBirthdayMessage } from "@/services/slack/constructAskBirthdayMessage";
 import { createSlackApp } from "@/services/slack/createSlackApp";
 import { getUserInfo } from "@/services/slack/getUserInfo";
-import { sendMockSqsMessage } from "@/testUtils/sendMockSqsMessage";
+import { sendMockSqsMessage } from "@/testUtils/unit/sendMockSqsMessage";
 
 const constants = vi.hoisted(() => ({
   otherBotUserId: "B999",
