+++ tests/unit/askPresentIdeasFromTeam.test.ts
@@ -1,6 +1,6 @@
-import "@/testUtils/mocks/mockDb";
-import "@/testUtils/mocks/mockEventBridge";
-import "@/testUtils/mocks/mockSlackApp";
+import "@/testUtils/unit/mockDb";
+import "@/testUtils/unit/mockEventBridge";
+import "@/testUtils/unit/mockSlackApp";
 
 import {
   EventBridgeClient,
@@ -21,9 +21,9 @@ import {
 import { users } from "@/db/schema";
 import type { Events } from "@/events";
 import { handler as askPresentIdeasFromTeam } from "@/functions/events/askPresentIdeasFromTeam";
-import { mockEventBridgePayload } from "@/testUtils/mocks/mockEventBridgePayload";
-import { sendMockSqsMessage } from "@/testUtils/sendMockSqsMessage";
 import { testDb } from "@/testUtils/testDb";
+import { mockEventBridgePayload } from "@/testUtils/unit/mockEventBridgePayload";
+import { sendMockSqsMessage } from "@/testUtils/unit/sendMockSqsMessage";
 
 dayjs.extend(utc);
