+++ tests/unit/memberJoinedChannel.test.ts
@@ -1,4 +1,4 @@
-import "@/testUtils/mocks/mockEventBridge";
+import "@/testUtils/unit/mockEventBridge";
 
 import {
   EventBridgeClient,
@@ -10,8 +10,8 @@ import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import type { Events } from "@/events";
 import { handler } from "@/functions/events/memberJoinedChannel";
 import { getUserInfo } from "@/services/slack/getUserInfo";
-import { mockEventBridgePayload } from "@/testUtils/mocks/mockEventBridgePayload";
-import { sendMockSqsMessage } from "@/testUtils/sendMockSqsMessage";
+import { mockEventBridgePayload } from "@/testUtils/unit/mockEventBridgePayload";
+import { sendMockSqsMessage } from "@/testUtils/unit/sendMockSqsMessage";
 
 const constants = vi.hoisted(() => ({
   birthdayBotUserId: "B001",
