--- tests/utils/callWithMockCronEvent.ts
@@ -1,11 +1,0 @@
-import { handler } from "@/functions/cron/daily";
-
-import { mockLambdaEvent } from "./mocks/mockLambdaPayload";
-
-export const callWithMockCronEvent = async (eventId: string) =>
-  handler({
-    ...mockLambdaEvent,
-    queryStringParameters: {
-      eventId,
-    },
-  });
