+++ tests/utils/constants.ts
@@ -1,2 +1,3 @@
 export const timeout = 20_000;
 export const waitTimeout = 18_000;
+export const pollInterval = 1_000;
